+++ contrib/platform/test/com/sun/jna/platform/win32/COM/ITypeInfoTest.java
@@ -25,9 +25,10 @@ import com.sun.jna.platform.win32.WinDef.WORDByReference;
 import com.sun.jna.platform.win32.WinNT.HRESULT;
 import com.sun.jna.ptr.IntByReference;
 import com.sun.jna.ptr.PointerByReference;
-import static junit.framework.TestCase.assertEquals;
 import org.junit.Ignore;
 import org.junit.Test;
+
+import static org.junit.Assert.*;
 
 /**
  * @author dblock[at]dblock[dot]org